// packages/sdk/src/router/v6/types.ts
export type ExecutionInfo = {
  module: string;
  data: string;
  value: BigNumberish;
};